zshrc: Fix tmux renaming window title to rename the correct window
A scenario that was broken before was when in window 1 you run `sleep 5`
then switch to window 2, the title of window 2 ends up being modified
when `sleep` completes.

This is because `tmux display -p '#{window_index}' is not an env
variable and local to the process but global value of the current window
in used. To workaround this, I'm creating an env variable so each
process have their own window id stored in MY_TMUX_WINDOW_ID.